--- Planilla_UCR/Application/Accounts/IAccountService.cs
@@ -1,11 +1,0 @@
-﻿using Domain.Accounts.Entities;
-using System.Threading.Tasks;
-
-namespace Application.Accounts
-{
-    public interface IAccountService
-    {
-        Task CreateAccountAsync(Account accountInfo);
-        Task InsertAccountData(Account accountData);
-    }
-}
