+++ Planilla_UCR/Application/DependencyInjection.cs
@@ -1,7 +1,5 @@
-﻿using Application.Persons;
-using Application.Accounts;
-using Application.Accounts.Implementations;
-using Application.Persons.Implementations;
+﻿using Application.People;
+using Application.People.Implementations;
 using Application.Employees;
 using Application.Employees.Implementations;
 using Microsoft.Extensions.DependencyInjection;
@@ -12,7 +10,6 @@ namespace Application
     {
         public static IServiceCollection AddApplicationLayer(this IServiceCollection services)
         {
-            services.AddTransient<IAccountService, AccountService>();
             services.AddTransient<IPersonService, PersonService>();
             services.AddTransient<IEmployeeService, EmployeeService>();
             return services;
