--- Planilla_UCR/Application/Employers/IEmployerService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-
-namespace Application.Employers
-{
-    public interface IEmployerService
-    {
-        Task CreateEmployerAsync(String email);
-    }
-}
