+++ Planilla_UCR/Application/People/IPersonService.cs
@@ -1,0 +1,13 @@
+﻿using Domain.People.Entities;
+using System.Threading.Tasks;
+using System.Collections.Generic;
+
+namespace Application.People
+{
+    public interface IPersonService
+    {
+        Task CreatePersonAsync(Person personInfo);
+
+        Task<IEnumerable<Person>> GetAllEmployees();
+    }
+}
