+++ Planilla_UCR/Application/People/Implementations/PersonService.cs
@@ -1,0 +1,27 @@
+﻿using Domain.People.Repositories;
+using Domain.People.Entities;
+using System.Threading.Tasks;
+using System.Collections.Generic;
+
+namespace Application.People.Implementations
+{
+    internal class PersonService : IPersonService
+    {
+        private readonly IPersonRepository _personRepository;
+
+        public PersonService(IPersonRepository personRepository)
+        {
+            _personRepository = personRepository;
+        }
+
+        public async Task CreatePersonAsync(Person personInfo)
+        {
+            await _personRepository.CreatePersonAsync(personInfo);
+        }
+
+        public async Task<IEnumerable<Person>> GetAllEmployees()
+        {
+            return await _personRepository.GetAllEmployees();
+        }
+    }
+}
