--- Planilla_UCR/Application/Persons/IPersonService.cs
@@ -1,14 +1,0 @@
-﻿using Domain.Persons.Entities;
-using Domain.Persons.DTOs;
-using System.Threading.Tasks;
-using System.Collections.Generic;
-
-namespace Application.Persons
-{
-    public interface IPersonService
-    {
-        Task CreatePersonAsync(Person personInfo);
-
-        Task<IEnumerable<Person>> GetAllEmployees();
-    }
-}
