--- Planilla_UCR/Domain/Accounts/Repositories/IAccountRepository.cs
@@ -1,11 +1,0 @@
-﻿using System.Threading.Tasks;
-using Domain.Accounts.Entities;
-
-namespace Domain.Accounts.Repositories
-{
-    public interface IAccountRepository
-    {
-        Task CreateAccountAsync(Account accountInfo);
-        Task InsertAccountData(Account accountData);
-    }
-}
