--- Planilla_UCR/Domain/Accounts/ValueObjects/AccountCode.cs
@@ -1,7 +1,0 @@
-﻿namespace Domain.Accounts.ValueObjects
-{
-    class AccountCode 
-    {
-
-    }
-}
