--- Planilla_UCR/Domain/Employers/Repositories/IEmployerRepository.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace Domain.Employers.Repositories
-{
-    public interface IEmployerRepository
-    {
-        Task CreateEmployerAsync(String email);
-    }
-}
