--- Planilla_UCR/Domain/Employers/ValueObjects/EmployerCode.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Domain.Employers.ValueObjects
-{
-    class EmployerCode 
-    {
-
-    }
-}
