+++ Planilla_UCR/Domain/People/DTOs/PersonDTO.cs
@@ -1,0 +1,35 @@
+﻿
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+
+namespace Domain.People.DTOs
+{
+    public class PersonDTO
+    {
+        public String Email { get; set; }
+        public String Name { get; set; }
+        public String LastName1 { get; set; }
+        public String LastName2 { get; set; }
+        public int Ssn { get; }
+        public String BankAccount { get; set; }
+        public String Adress { get; set; }
+        public String PhoneNumber { get; set; }
+
+        public PersonDTO(String email, String name, String lastName1, String lastName2, int id, String bankAccount,
+                          String adress, String phoneNumber)
+        {
+            Email = email;
+            Name = name;
+            LastName1 = lastName1;
+            LastName2 = lastName2;
+            Ssn = id;
+            BankAccount = bankAccount;
+            Adress = adress;
+            PhoneNumber = phoneNumber;
+        }
+    }
+}
