--- Planilla_UCR/Infrastructure/Accounts/AccountDbContext.cs
@@ -1,23 +1,0 @@
-﻿using Domain.Accounts.Entities;
-using Infrastructure.Core;
-using Infrastructure.Accounts.EntityMappings;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.Extensions.Logging;
-
-namespace Infrastructure.Accounts
-{
-    public class AccountDbContext : ApplicationDbContext
-    {
-        public AccountDbContext(DbContextOptions<AccountDbContext> options, ILogger<AccountDbContext> logger) : base(options, logger)
-        {
-        }
-
-        public DbSet<Account> Accounts { get; set; } = null!;
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder);
-            modelBuilder.ApplyConfiguration(new AccountMap());
-        }
-    }
-}
