--- Planilla_UCR/Infrastructure/Accounts/EntityMappings/AccountMap.cs
@@ -1,21 +1,0 @@
-﻿using Domain.Accounts.Entities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-
-namespace Infrastructure.Accounts.EntityMappings
-{
-    public class AccountMap : IEntityTypeConfiguration<Account>
-    {
-        public void Configure(EntityTypeBuilder<Account> builder)
-        {
-            builder.ToTable("Account");
-
-            builder.HasKey(p => p.Email);
-
-            builder.Property(p => p.Password)
-                   .IsRequired();
-        }
-
-    }
-}
