+++ Planilla_UCR/Infrastructure/DependencyInjection.cs
@@ -1,11 +1,6 @@
-﻿using Infrastructure.Persons;
-using Domain.Core.Repositories;
-using Infrastructure.Accounts;
-using Infrastructure.Accounts.Repositories;
-using Domain.Accounts.Repositories;
-using Infrastructure.Persons;
-using Infrastructure.Persons.Repositories;
-using Domain.Persons.Repositories;
+﻿using Infrastructure.People;
+using Infrastructure.People.Repositories;
+using Domain.People.Repositories;
 
 using Infrastructure.Employees;
 using Infrastructure.Employees.Repositories;
@@ -20,8 +15,6 @@ namespace Infrastructure
     {
         public static IServiceCollection AddInfrastructureLayer(this IServiceCollection services, string connectionString)
         {
-            services.AddDbContext<AccountDbContext>(options => options.UseSqlServer(connectionString));
-            services.AddScoped<IAccountRepository, AccountRepository>();
 
             services.AddDbContext<PersonDbContext>(options => options.UseSqlServer(connectionString));
             services.AddScoped<IPersonRepository, PersonRepository>();
