--- Planilla_UCR/Presentation/Accounts/Components/Employee2Form.razor
@@ -1,178 +1,0 @@
-﻿@page "/Create_Employe2"
-@using Application.Persons
-@using Domain.Persons.Entities
-
-@using Application.Employees
-@using Domain.Employees.Entities
-
-@using Presentation.Core.Components
-@using Presentation.Employees.Models
-@inject ISnackbar Snackbar
-@inject IPersonService PersonService
-@inject IEmployeeService EmployeeService
-
-@*Navigation footprint *@
-@{
-    List<BreadcrumbItem> _items = new List<BreadcrumbItem>
-    {
-        new BreadcrumbItem("Inicio", href: "#"),
-        new BreadcrumbItem("Cuenta", href: "#"),
-    };
-}
- <MudBreadcrumbs Items="_items" Separator=">"></MudBreadcrumbs>
-
-<MudGrid>
-    <MudItem xs="12" sm="8">
-        <MudPaper  Height="220px" Width="90%" Outlined="false" Class="border-none pa-4" Elevation="0">
-            <MudText Align="Align.Center" Typo="Typo.h6">Información personal</MudText>
-            <MudForm @ref="form">
-                    <MudGrid>
-                    @*Small fields *@
-                    <MudItem xs="6">
-                        <MudTextField @bind-Value="personInfo.Name" Label="Nombre" Variant="Variant.Outlined" Required="true" RequiredError="Campo requerido"></MudTextField>
-                    </MudItem>
-
-                    <MudItem xs="6">
-                        <MudTextField @bind-Value="personInfo.Email" Label="Email" Variant="Variant.Outlined" Required="true" RequiredError="Campo requerido"></MudTextField>
-                    </MudItem>
-
-                    <MudItem xs="6">
-                        <MudTextField @bind-Value="personInfo.LastName1" Label="Primer Apellido" Variant="Variant.Outlined"></MudTextField>
-                    </MudItem>
-
-                    <MudItem xs="6">
-                        <MudTextField @bind-Value="personInfo.LastName2" Label="Segundo Apellido" Variant="Variant.Outlined"></MudTextField>
-                    </MudItem>
-
-                     <MudItem xs="6">
-                        <MudTextField @bind-Value="personInfo.PhoneNumber" Label="Telefono" Variant="Variant.Outlined" ></MudTextField>
-                    </MudItem>
-
-                    <MudItem xs="6">
-                        <MudTextField @bind-Value="personInfo.Ssn" Label="Cedula" Variant="Variant.Outlined" Required="true" RequiredError="Campo requerido"></MudTextField>
-                    </MudItem>
-
-                    @*Large fields *@
-                    <MudItem xs="12">
-                        <MudTextField @bind-Value="personInfo.Adress" Label="Dirección" Variant="Variant.Outlined"></MudTextField>
-                    </MudItem>
-
-                    <MudItem xs="12">
-                        <MudTextField @bind-Value="personInfo.BankAccount" Label="Cuenta Bancaria" Variant="Variant.Outlined" Required="true" RequiredError="Campo requerido"></MudTextField>
-                    <MudDivider />
-                    </MudItem>
-                
-
-                    @*Create buton *@
-                    <MudItem xs="12" sm="6" md="4">
-                        <MudButton Disabled="@snackHandler._saving" OnClick="SavePersonInfo" Variant="Variant.Filled" Color="Color.Secondary">
-                            @if (snackHandler._saving)
-                            {
-                                <MudProgressCircular Class="ms-n1" Size="Size.Small" Indeterminate="true"/>
-                                <MudText Class="ms-2">Procesando</MudText>
-                            }
-                            else
-                            {
-                                <MudText>Crear</MudText>
-                            }
-                        </MudButton>
-                    </MudItem>
-
-                     @*Clear buton *@
-                    <MudItem xs="12" sm="6" md="4">
-                        <MudButton OnClick="CleanForm" Variant="Variant.Filled" Color="Color.Error">
-                                <MudText>Limpiar</MudText>
-                        </MudButton>
-                    </MudItem>
-
-                </MudGrid>
-            </MudForm>
-        </MudPaper>
-    </MudItem>
-
-    @*File uploader *@
-    <MudItem xs="3">
-        <FileUploader>
-        </FileUploader>   
-    </MudItem>
-
-</MudGrid>
-
-@code {
-    Person personInfo = new Person("", "", "", "", Int32.Parse("0"), "", "", "");
-    SnackHandler snackHandler = new SnackHandler();
-    MudForm form;
-
-    async Task SavePersonInfo()
-    {
-        snackHandler._saving = true;
-        await form.Validate();
-        if (form.IsValid && ValidateInfo(personInfo))
-        {
-            await PersonService.CreatePersonAsync(personInfo);
-            await EmployeeService.CreateEmployeeAsync(personInfo.Email);
-            snackHandler._saving = false;
-            ShowSnackNotification("Empleado "+personInfo.Name+" creado", Severity.Success);
-        }
-        snackHandler._saving = false;
-    }
-
-    bool ValidateSsn(Person personInfo)
-    {
-        bool valid = true;
-        if (personInfo.Ssn == 0 || personInfo.Ssn < 0)
-        {
-            ShowSnackNotification("Número de cédula incorrecta", Severity.Error);
-            valid = false;
-        }
-        return valid;
-    }
-
-    bool ValidateEmail(Person personInfo)
-    {
-        bool valid = true;
-
-        valid = personInfo.Email.Contains('@');
-        if (!valid)
-        {
-            ShowSnackNotification("El email debe contener un @", Severity.Error);
-        }
-        return valid;
-    }
-
-    bool ValidateInfo(Person personInfo)
-    {
-        bool ssn = ValidateSsn(personInfo);
-        bool email = ValidateEmail(personInfo);
-        return ssn && email;
-    }
-
-    void ShowSnackNotification(string message, Severity severity)
-    {
-        Snackbar.Configuration.SnackbarVariant = Variant.Filled;
-        Snackbar.Configuration.PreventDuplicates = false;
-        Snackbar.Configuration.MaxDisplayedSnackbars = 10;
-        Snackbar.Configuration.NewestOnTop = true;
-        Snackbar.Add($"{message}", severity, config =>
-        {
-            config.RequireInteraction = false;
-            config.ShowCloseIcon = true;
-            config.Onclick = snackbar =>
-            {
-                return Task.CompletedTask;
-            };
-        });
-    }
-
-    void CleanForm()
-    {
-        personInfo.Name = "";
-        personInfo.Email = "";
-        personInfo.LastName1 = "";
-        personInfo.LastName2 = "";
-        personInfo.Adress = "";
-        personInfo.BankAccount = "";
-        personInfo.PhoneNumber = "";
-        personInfo.Ssn = 0;
-    }
-}
