--- Planilla_UCR/Presentation/Accounts/Components/FileUploader.razor
@@ -1,64 +1,0 @@
-﻿@inject ISnackbar Snackbar
-
-<br />
-<MudText Align="Align.Center" Typo="Typo.h6">Foto de perfil</MudText>
-<MudPaper @ondragenter="@SetDragClass" @ondragleave="@ClearDragClass" @ondragend="@ClearDragClass" Height="150px" Outlined="true" Class="@DragClass">
-    @if (!Clearing)
-    {
-        <InputFile OnChange="OnInputFileChanged" multiple  class="absolute mud-width-full mud-height-full overflow-hidden z-2" style="opacity:0;" />
-    }
-    <MudText Typo="Typo.body1">Arrastra la imagen, o haz click aquí</MudText>
-    @foreach (var file in fileNames)
-    {
-        <MudChip Color="Color.Dark" Text="@file" />
-    }
-</MudPaper>
-<MudToolBar DisableGutters="true" Class="gap-4">
- <MudSpacer></MudSpacer>
-<MudButton OnClick="Upload" Disabled="@(!fileNames.Any())" Color="Color.Secondary" Variant="Variant.Filled" class= "align-center">Subir</MudButton>
-<MudSpacer></MudSpacer>
-<MudButton OnClick="Clear" Disabled="@(!fileNames.Any())" Color="Color.Error" Variant="Variant.Filled" class= "align-center">Limpiar</MudButton>
-<MudSpacer></MudSpacer>
-</MudToolBar>
-
-@code {
-    private bool Clearing = false;
-    private static string DefaultDragClass = "relative rounded-lg border-2 border-dashed pa-4 mt-4 mud-width-full mud-height-full";
-    private string DragClass = DefaultDragClass;
-    private List<string> fileNames = new List<string>();
-
-    private void OnInputFileChanged(InputFileChangeEventArgs e)
-    {
-        ClearDragClass();
-        var files = e.GetMultipleFiles();
-        foreach(var file in files)
-        {
-            fileNames.Add(file.Name);
-        }
-    }
-
-    private async Task Clear()
-    {
-        Clearing = true;
-        fileNames.Clear();
-        ClearDragClass();
-        await Task.Delay(0);
-        Clearing = false;
-    }
-    private void Upload()
-    {
-        //Upload the files here
-        Snackbar.Configuration.PositionClass = Defaults.Classes.Position.TopCenter;
-        Snackbar.Add("TODO: Upload your files!", Severity.Normal);
-    }
-
-    private void SetDragClass()
-    {
-        DragClass = $"{DefaultDragClass} mud-border-primary";
-    }
-
-    private void ClearDragClass()
-    {
-        DragClass = DefaultDragClass;
-    }
-}
