--- Planilla_UCR/Presentation/Accounts/Models/SnackHandler.cs
@@ -1,17 +1,0 @@
-﻿
-namespace Presentation.Accounts.Models
-{
-    internal class SnackHandler
-    {
-        public bool _saving { get; set; }
-        public bool _validInfo { get; set; }
-        public bool _created { get; set; }
-
-        public SnackHandler()
-        {
-            _saving = false;
-            _validInfo = false;
-            _created = false;
-        }
-    }
-}
