+++ Planilla_UCR/Presentation/Agreements/Index.razor
@@ -1,0 +1,7 @@
+﻿@page "/Agreements/{employeeEmail}"
+
+    <p>esto @employeeEmail</p>
+@code {
+    [Parameter]
+    public string employeeEmail { get; set; }
+}
