+++ Planilla_UCR/Presentation/Core/Layout/NavMenu.razor
@@ -7,15 +7,6 @@
         </li>
 
         <li class="nav-item px-3">
-             <NavLink class="nav-link" href="Loggin" Match="NavLinkMatch.Prefix">
-                <span class="fas fa-hockey-puck pr-3" aria-hidden="true"></span>
-                Iniciar Sesión
-            </NavLink>
-            <NavLink class="nav-link" href="Register_Form" Match="NavLinkMatch.Prefix">
-                <span class="fas fa-hockey-puck pr-3" aria-hidden="true"></span>
-                Registrarse
-            </NavLink>
-            
             <NavLink class="nav-link" href="Create_Employee" Match="NavLinkMatch.Prefix">
                 <span class="fas fa-hockey-puck pr-3" aria-hidden="true"></span>
                 Crear Empleado
@@ -23,10 +14,6 @@
             <NavLink class="nav-link" href="Hire_Employees" Match="NavLinkMatch.Prefix">
                 <span class="fas fa-hockey-puck pr-3" aria-hidden="true"></span>
                 Contratar empleados 
-            </NavLink>
-            <NavLink class="nav-link" href="View_Hired_Employees" Match="NavLinkMatch.Prefix">
-                <span class="fas fa-hockey-puck pr-3" aria-hidden="true"></span>
-                Empleados Contratados
             </NavLink>
         </li>
