+++ Planilla_UCR/Presentation/Employees/Pages/CreateEmployee.razor
@@ -1,17 +1,17 @@
 ﻿@page "/Create_Employee"
 
-@using Application.Persons
-@using Domain.Persons.Entities
-@using Application.Accounts
-@using Domain.Accounts.Entities
-@using Presentation.Accounts.Models
+@using Application.People
+@using Domain.People.Entities
+@*@using Application.Accounts
+@using Domain.Accounts.Entities*@
+@*@using Presentation.Accounts.Models*@
 @using Application.Employees
+@using Presentation.Employees.Models
 @using Presentation.Core.Components
 @using System.ComponentModel.DataAnnotations
 
 @inject ISnackbar Snackbar
 @inject IPersonService PersonService
-@inject IAccountService AccountService
 @inject IEmployeeService EmployeeService
 
 @*Navigation footprint *@
@@ -109,7 +109,7 @@
 
 @code {
     Person personInfo = new Person("", "", "", "", Int32.Parse("0"), "", "", "");
-    Account accountInfo = new Account("","");
+    //Account accountInfo = new Account("","");
     SnackHandler snackHandler = new SnackHandler();
     MudForm form;
     
@@ -120,9 +120,9 @@
         if (form.IsValid && ValidateSsn(personInfo))
         {
             await PersonService.CreatePersonAsync(personInfo);
-            accountInfo.Email = personInfo.Email;
-            accountInfo.Password = personInfo.Name +"_"+ personInfo.Ssn.ToString();
-            await AccountService.CreateAccountAsync(accountInfo);
+            //accountInfo.Email = personInfo.Email;
+            //accountInfo.Password = personInfo.Name +"_"+ personInfo.Ssn.ToString();
+            //await AccountService.CreateAccountAsync(accountInfo);
             await EmployeeService.CreateEmployeeAsync(personInfo.Email);
             snackHandler._saving = false;
             ShowSnackNotification("Cuenta para empleado "+personInfo.Name+" creada", Severity.Success);
@@ -170,6 +170,6 @@
         personInfo.BankAccount = "";
         personInfo.PhoneNumber = "";
         personInfo.Ssn = 0;
-        accountInfo.Password = "";
+        //accountInfo.Password = "";
     }
 }
