--- Planilla_UCR/Presentation/Employer/Components/ViewProjectEmployeesError.razor
@@ -1,39 +1,0 @@
-﻿@using Application.Employees
-@using Domain.Employees.Entities
-@using Presentation.Core.Components
-@using Presentation.Employer
-@inject NavigationManager nav
-@inject IEmployeeService EmployeeServic
-
-@{
-    List<BreadcrumbItem> _items = new List<BreadcrumbItem>
-    {
-        new BreadcrumbItem("Proyectos", href: "#"),
-        new BreadcrumbItem("Empleados", href: "#"),
-    };
-}
-    <MudBreadcrumbs Items="_items" Separator=">"></MudBreadcrumbs>
-    <br />
-    <MudText Align="Align.Center" Typo="Typo.h4">No hay empleados contratados</MudText>
-    <br />
-    <br />
- 
-   <MudToolBar DisableGutters="true" Class="gap-4">
-        <MudButton></MudButton>
-        <MudSpacer></MudSpacer>
-        @*Create button *@
-        <MudButton Variant="Variant.Filled" OnClick="GoToCreateEmployee" Color="Color.Secondary">
-            <MudText> 
-            Contratae empleados
-            </MudText> 
-        </MudButton>
-        <MudSpacer></MudSpacer>
-        <MudButton></MudButton>
-    </MudToolBar>
-
-@code {
-   private void GoToCreateEmployee()
-    {
-        nav.NavigateTo("/Asociate_Employees");
-    }
-}
