+++ Planilla_UCR/Presentation/Employers/Components/HireEmployeesError.razor
@@ -1,0 +1,40 @@
+﻿@using Application.Employees
+@using Domain.Employees.Entities
+@using Presentation.Core.Components
+@inject NavigationManager nav
+@inject IEmployeeService EmployeeServic
+
+@*Navigation footprint *@
+@{
+    List<BreadcrumbItem> _items = new List<BreadcrumbItem>
+    {
+        new BreadcrumbItem("Proyectos", href: "#"),
+        new BreadcrumbItem("Empleados", href: "#"),
+    };
+}
+    <MudBreadcrumbs Items="_items" Separator=">"></MudBreadcrumbs>
+    <br />
+    <MudText Align="Align.Center" Typo="Typo.h4">No hay empleados registrados</MudText>
+    <br />
+    <br />
+ 
+   <MudToolBar DisableGutters="true" Class="gap-4">
+        <MudButton></MudButton>
+        <MudSpacer></MudSpacer>
+        @*Create button *@
+        <MudButton Variant="Variant.Filled" OnClick="GoToCreateEmployee" Color="Color.Secondary">
+            <MudText> 
+            Crear empleado
+            </MudText> 
+        </MudButton>
+        <MudSpacer></MudSpacer>
+        <MudButton></MudButton>
+    </MudToolBar>
+
+@code {
+
+   private void GoToCreateEmployee()
+    {
+        nav.NavigateTo("/Create_Employee");
+    }
+}
