+++ Planilla_UCR/Presentation/Employers/Pages/HireEmployees.razor
@@ -1,0 +1,119 @@
+﻿@page "/Hire_Employees"
+@using Application.Employees
+@using Application.People
+@using Domain.Employees.Entities
+@using Domain.Employees.DTOs
+@using Domain.People.Entities
+@using Presentation.Employers.Components
+@using Presentation.Core.Components
+@inject NavigationManager nav
+@inject IEmployeeService EmployeeService
+@inject IPersonService PersonService
+
+@if(_loading) {
+
+    <MudPaper  Height="220px" Width="100%" Outlined="false" Class="border-none pa-4" Elevation="0">
+         <MudText Align="Align.Center" Typo="Typo.h6">
+              <MudProgressCircular Color="Color.Secondary" Style="height:70px;width:70px;" Indeterminate="true" />
+         </MudText>
+       
+    </MudPaper>
+}else{
+    @if (!empty && !_loading)
+    {
+        List<BreadcrumbItem> _items = new List<BreadcrumbItem>
+        {
+            new BreadcrumbItem("Proyectos", href: "#"),
+            new BreadcrumbItem("Empleados", href: "#"),
+        };
+
+        <MudBreadcrumbs Items="_items" Separator=">"></MudBreadcrumbs>
+
+        <MudPaper  Height="220px" Width="100%" Outlined="false" Class="border-none pa-4" Elevation="0">
+            <MudTable @ref="@_table" Items="@_elements" RowsPerPage="4" Hover="true" Breakpoint="Breakpoint.Sm" Loading="@_loading" Filter="new Func<Person,bool>(FilterFunc1)">
+                    <ToolBarContent>
+                         <MudText Align="Align.Center" Typo="Typo.h6">Contratar empleados</MudText>
+                        <MudSpacer />
+                        <MudTextField @bind-Value="searchString1" Placeholder="Search" Adornment="Adornment.Start" AdornmentIcon="@Icons.Material.Filled.Search" IconSize="Size.Medium" Class="mt-0"></MudTextField>
+                    </ToolBarContent>
+                    <HeaderContent>
+                        <MudTh>Email</MudTh>
+                        <MudTh>Nombre completo</MudTh>
+                        <MudTh>Cedula</MudTh>
+                        <MudTh></MudTh>
+                    </HeaderContent>
+
+                    <RowTemplate>
+                        <MudTd DataLabel="Email">@context.Email</MudTd>
+                        <MudTd DataLabel="Nombre">@context.Name @context.LastName1 @context.LastName2</MudTd>
+                        <MudTd DataLabel="Cedula">@context.Ssn</MudTd>
+                         <MudTd DataLabel="">
+                        <MudSpacer></MudSpacer>
+                            <MudButton Variant="Variant.Filled" @onclick="@(e => GoToContractEmployee(@context.Email))" Color="Color.Secondary">
+                                <MudText>
+                                Contratar
+                                </MudText> 
+                            </MudButton>
+                        </MudTd>
+                    </RowTemplate>
+                <PagerContent>
+                    <MudPagination SelectedChanged="PageChanged" Count="@((_table.GetFilteredItemsCount() + _table.RowsPerPage - 1) / _table.RowsPerPage)" Color="Color.Secondary" Class="pa-4"/>
+                </PagerContent>
+            </MudTable>
+        </MudPaper>
+    }
+    else
+    {
+        <HireEmployeesError></HireEmployeesError>
+    }
+}
+
+@code {
+    private bool _hidePosition;
+    private bool _loading = true;
+    private bool empty = true;
+    string searchString1 = "";
+    private IEnumerable<Person> _elements = new List<Person>();
+    private MudTable<Person> _table;
+
+    protected override async Task OnInitializedAsync()
+    {
+
+        _elements = await PersonService.GetAllEmployees();
+        if (_elements.Length() == 0)
+        {
+            empty = true;
+        }
+        else
+        {
+            empty = false;
+        }
+        _loading = false;
+    }
+
+    private void GoToCreateEmployee()
+    {
+        nav.NavigateTo("/Create_Employee");
+    }
+
+    private void GoToContractEmployee(String email)
+    {
+       nav.NavigateTo($"/Agreements/{email}");
+    }
+
+    private void PageChanged(int i)
+    {
+        _table.NavigateTo(i - 1);
+    }
+
+    private bool FilterFunc1(Person element) => FilterFunc(element, searchString1);
+
+    private bool FilterFunc(Person element, string searchString)
+    {
+        if (string.IsNullOrWhiteSpace(searchString))
+            return true;
+        if (element.Email.Contains(searchString, StringComparison.OrdinalIgnoreCase))
+            return true;
+        return false;
+    }
+}
